# docs/swagger.yaml
basePath: /
definitions:
  models.Comment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo:
        $ref: '#/definitions/models.Photo'
      photoID:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.CommentCreate:
    properties:
      message:
        type: string
    type: object
  models.Photo:
    properties:
      caption:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.PhotoCreate:
    properties:
      caption:
        type: string
      photo_url:
        type: string
      title:
        type: string
    type: object
  models.SocialMedia:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.SocialMediaCreate:
    properties:
      name:
        type: string
      social_media_url:
        type: string
    type: object
  models.User:
    properties:
      age:
        type: integer
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  models.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.UserRegister:
    properties:
      age:
        type: integer
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a photo sharing app.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: My-Garm API Documentation
  version: 1.0.0
paths:
  /comments/{photoId}:
    get:
      consumes:
      - application/json
      description: Get Details of Comments by corresponding Photo ID
      parameters:
      - description: photo id
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
      security:
      - Bearer: []
      summary: Get Comments
      tags:
      - Comment
    post:
      consumes:
      - application/json
      description: Create New Comment Based on Photo ID
      parameters:
      - description: photo id
        in: path
        name: photoId
        required: true
        type: integer
      - description: create comment
        in: body
        name: CommentCreate
        required: true
        schema:
          $ref: '#/definitions/models.CommentCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
      security:
      - Bearer: []
      summary: Create Comment
      tags:
      - Comment
  /comments/{photoId}/{commentId}:
    delete:
      consumes:
      - application/json
      description: Delete Comment by corresponding ID
      parameters:
      - description: photo id
        in: path
        name: photoId
        required: true
        type: integer
      - description: comment id
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
      security:
      - Bearer: []
      summary: Delete Comment
      tags:
      - Comment
    get:
      consumes:
      - application/json
      description: Get Details of Comment by corresponding ID
      parameters:
      - description: photo id
        in: path
        name: photoId
        required: true
        type: integer
      - description: comment id
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
      security:
      - Bearer: []
      summary: Get Comment
      tags:
      - Comment
    put:
      consumes:
      - application/json
      description: Update Comment by corresponding ID
      parameters:
      - description: photo id
        in: path
        name: photoId
        required: true
        type: integer
      - description: comment id
        in: path
        name: commentId
        required: true
        type: integer
      - description: update comment
        in: body
        name: CommentUpdate
        required: true
        schema:
          $ref: '#/definitions/models.CommentCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
      security:
      - Bearer: []
      summary: Update Comment
      tags:
      - Comment
  /photos:
    get:
      consumes:
      - application/json
      description: Get Details of All Photos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Photo'
      security:
      - Bearer: []
      summary: Get Photos
      tags:
      - Photo
    post:
      consumes:
      - application/json
      description: Create New Photo
      parameters:
      - description: create photo
        in: body
        name: PhotoCreate
        required: true
        schema:
          $ref: '#/definitions/models.PhotoCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Photo'
      security:
      - Bearer: []
      summary: Create Photo
      tags:
      - Photo
  /photos/{photoId}:
    delete:
      consumes:
      - application/json
      description: Delete Photo by corresponding ID
      parameters:
      - description: Photo ID
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Photo'
      security:
      - Bearer: []
      summary: Delete Photo
      tags:
      - Photo
    get:
      consumes:
      - application/json
      description: Get Details of Photo by corresponding ID
      parameters:
      - description: Photo ID
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Photo'
      security:
      - Bearer: []
      summary: Get Photo
      tags:
      - Photo
    put:
      consumes:
      - application/json
      description: Update Details of Photo by corresponding ID
      parameters:
      - description: Photo ID
        in: path
        name: photoId
        required: true
        type: integer
      - description: update photo
        in: body
        name: PhotoUpdate
        required: true
        schema:
          $ref: '#/definitions/models.PhotoCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Photo'
      security:
      - Bearer: []
      summary: Update Photo
      tags:
      - Photo
  /social-media:
    get:
      consumes:
      - application/json
      description: Get Details of all Social Medias
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SocialMedia'
      security:
      - Bearer: []
      summary: Get Social Medias
      tags:
      - Social Media
    post:
      consumes:
      - application/json
      description: Create New Social Media
      parameters:
      - description: create social media
        in: body
        name: SocialMediaCreate
        required: true
        schema:
          $ref: '#/definitions/models.SocialMediaCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SocialMedia'
      security:
      - Bearer: []
      summary: Create Social Media
      tags:
      - Social Media
  /social-media/{socialMediaId}:
    delete:
      consumes:
      - application/json
      description: Delete Social Media corresponding to socialMediaId
      parameters:
      - description: Social Media Id
        in: path
        name: socialMediaId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SocialMedia'
      security:
      - Bearer: []
      summary: Delete Social Media
      tags:
      - Social Media
    get:
      consumes:
      - application/json
      description: Get Details of Social Media corresponding to socialMediaId
      parameters:
      - description: Social Media Id
        in: path
        name: socialMediaId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SocialMedia'
      security:
      - Bearer: []
      summary: Get Social Media
      tags:
      - Social Media
    put:
      consumes:
      - application/json
      description: Update Details of Social Media corresponding to socialMediaId
      parameters:
      - description: Social Media Id
        in: path
        name: socialMediaId
        required: true
        type: integer
      - description: update social media
        in: body
        name: SocialMedia
        required: true
        schema:
          $ref: '#/definitions/models.SocialMediaCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SocialMedia'
      security:
      - Bearer: []
      summary: Update Social Media
      tags:
      - Social Media
  /users/login:
    post:
      consumes:
      - application/json
      description: Login user
      operationId: login-user
      parameters:
      - description: email
        in: body
        name: UserLogin
        required: true
        schema:
          $ref: '#/definitions/models.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
      summary: Login user
      tags:
      - User
  /users/register:
    post:
      consumes:
      - application/json
      description: Register new user
      operationId: register-user
      parameters:
      - description: UserRegister
        in: body
        name: UserRegister
        required: true
        schema:
          $ref: '#/definitions/models.UserRegister'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Register new user
      tags:
      - User
securityDefinitions:
  Bearer:
    description: Type "Bearer" eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2ODE0ODA2MDUsImlzcyI6Ik1ZLUdBUk0iLCJpZCI6NCwiRW1haWwiOiJva0Bvay5jb20ifQ.bPcX66MlkvCWypEXpLEadFToWYDoe8w3iq9YXp_JlHs
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"